test(*): Clear overlays after test execution, #2029
