+++ projects/igniteui-angular/src/lib/test-utils/helper-utils.spec.ts
@@ -1,0 +1,11 @@
+export class HelperUtils {
+    
+    public static clearOverlay() {
+        const overlays = document.getElementsByClassName('igx-overlay') as HTMLCollectionOf<Element>;
+        Array.from(overlays).forEach(element => {
+            element.remove();
+        });
+        document.documentElement.scrollTop = 0;
+        document.documentElement.scrollLeft = 0;
+    }
+}
